Fixed catalog update issues in worker
